lstm/requirements.txt
streamlit
pandas
yfinance
tensorflow-cpu
scikit-learn
plotly
joblib